# message-mesh-svc/src/main/resources/message-mesh-openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Message Mesh API",
...
    "contact": {
      "name": "Message Mesh Platform Team",
      "url": "https://message-mesh.dev.com",
      "email": "platform@message-mesh.dev.com"
    },
    "license": {
      "name": "Apache License 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.message-mesh.prod.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/conversations": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "summary": "List the caller's conversations",
        "description": "Returns every conversation the authenticated user is a member of, each enriched with its most recent message and the caller's unread count. Useful for rendering the conversation sidebar.",
        "operationId": "list_1",
        "responses": {
          "200": {
            "description": "Conversations returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Create a conversation",
        "description": "Creates a DIRECT (exactly two members) or GROUP conversation. The caller is added automatically; 'memberUsernames' lists the other participants. For DIRECT chats an existing conversation with the same pair may be reused.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "One or more member usernames do not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "201": {
            "description": "Conversation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload (e.g. empty members, bad type)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{id}/read": {
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Mark a conversation as read",
        "description": "Advances the caller's read marker to the supplied sequence number, clearing the unread count up to and including that message and triggering read receipts to other members. Returns 204 No Content on success.",
        "operationId": "markRead",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Read marker updated"
          },
          "400": {
            "description": "Sequence must be zero or positive",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not a member of the conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{id}/members": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "summary": "List a conversation's members",
        "description": "Returns every participant of the conversation together with their membership role (ADMIN or MEMBER). Useful for rendering a group roster showing the member count and who the admins are. Admins are returned first.",
        "operationId": "members",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Caller is not a member of the conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Members returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationMemberDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Add members to a group",
        "description": "Adds one or more users (by username) to a GROUP conversation. Only group admins may add.",
        "operationId": "addMembers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMembersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Not a group or caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "A username does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Members added; updated roster returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationMemberDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations/{id}": {
      "delete": {
        "tags": [
          "Conversations"
        ],
        "summary": "Delete or leave a conversation",
        "description": "A group admin soft-deletes the whole conversation for everyone; any other member simply leaves (their membership is removed).",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Caller is not a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204": {
            "description": "Conversation deleted or left"
          }
        }
      },
      "patch": {
        "tags": [
          "Conversations"
        ],
        "summary": "Rename a group conversation",
        "description": "Updates the title of a GROUP conversation. Only group admins may rename.",
        "operationId": "rename",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Conversation renamed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "400": {
            "description": "Not a group, not an admin, or invalid title",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{id}/membership": {
      "patch": {
        "tags": [
          "Conversations"
        ],
        "summary": "Update per-user conversation preferences",
        "description": "Toggles the caller's own mute and/or archive flags for the conversation. Omitted fields are left unchanged.",
        "operationId": "updateMembership",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipPrefsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Caller is not a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Preferences updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{id}/members/{userId}/role": {
      "patch": {
        "tags": [
          "Conversations"
        ],
        "summary": "Change a member's role",
        "description": "Promotes or demotes a group member. Only group admins may change roles; the last admin cannot be demoted.",
        "operationId": "updateMemberRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Conversation identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "Target user identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Role updated; updated roster returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationMemberDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Not permitted or would demote the last admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/conversations/{id}/messages": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List messages in a conversation",
        "description": "Returns a page of messages (newest first) for moderation review, including soft-deleted tombstones. 'page' (0-based) and 'size' (max 100) paginate.",
        "operationId": "messages_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Caller is not an administrator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Messages returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseMessageDto"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "DIRECT",
              "GROUP"
            ]
          },
          "title": {
            "type": "string"
          },
          "lastMessage": {
            "$ref": "#/components/schemas/MessageDto"
          },
          "unreadCount": {
            "type": "integer",
            "format": "int64"
          },
          "memberCount": {
            "type": "integer",
            "format": "int32"
          },
          "muted": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
      "ReadRequest": {
        "type": "object",
        "properties": {
          "seq": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AddMembersRequest": {
        "required": [
          "usernames"
        ],
        "type": "object",
        "properties": {
          "usernames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT bearer token obtained from the authentication endpoints. Paste only the raw token value; Swagger UI adds the 'Bearer ' prefix.",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}